Move typedb-visualiser source code into this repo (#167)
## What is the goal of this PR?

Originally, the TypeDB graph visualisations on vaticle.com were powered by `typedb-visualiser`, an external library whose source code lived in the `typedb-studio` repo. It did so because we were planning to build Studio in Electron, with the visualiser code hence being reusable across the two repos.

But since we changed Studio to use Compose and Kotlin, the library lost its value as a standalone library; it's now only used by `web-main`. So we've absorbed its entire source code into this repo, and deleted the scripts that were previously used to manually publish the package (the publish process had not been automated - it had to be run on your local machine to deploy the package to npm)

## What are the changes made in this PR?

- Extract from `typedb-studio` the source code of the `typedb-visualiser` npm package, which was previously used by `web-main`, and move the code into `web-main` itself, eliminating the external dependency