Use MatIcon to display svg icons (#255)
Previously there were multiple solutions used to display icons.
Now MatIcon with custom svg icons is used where possible.

A few solutions were used to display icons:
- `img` tag with src pointing to `svg` image - this is not the perfect
solution as we can't control colour of the icon easly,
- inline svg - that would require us to create component for each icon
or copy svg code that is hard to maintain.

Using MatIcon makes it easier to control color and size as we just
adjust CSS `font-size` and `color`.

Removed all icon assets that are not accessible in the application.