Handle rich text cleanly (#283)
## What is the goal of this PR?

Sanity Portable Text is no longer rendered through a converter class and
component that manually parses each Portable Text tag, but rather
through a direct Portable Text to HTML converter (library maintained by
Sanity themselves). This allows us to support all features of Portable
Text.

## What are the changes implemented in this PR?

When we originally wrote Portable Text support, we identified a subset
of the most important features, then proceeded to the shortest path to
supporting them. This was to write a `RichText` class whose purpose was
to take the raw Portable Text data from Sanity and convert it into an
array of `span` elements, each with optional formatting (`marks`).

The limitations of this approach rapidly became obvious and grew in
severity. Whenever we wanted to support a new kind of formatting (e.g.
italics) we'd have to manually add code to do that to our own converter.
Rendering bulleted lists was hacky because they actually come back in a
different format from Sanity - not fully representable by a list of
spans.

Now, we've ripped all that up and rebuilt the rich text renderer using a
Portable Text to HTML converter (library maintained by Sanity
themselves). We introduce an `HtmlPipe` that takes PortableText and
pipes out HTML using this library. This allows us to dissolve both the
`RichText` class and the `td-rich-text` Angular component.