Upgrade to Angular 17 (#317)
## Release notes: usage and product changes

We've upgraded from Angular 16 to 17. Builds are now significantly
faster and we can utilise the latest Angular features.

## Implementation

- The project now uses Angular's `esbuild` based builder which builds
the website over 2x as fast as the previous one. The output bundle size
is also drastically reduced which improves performance for users.
- We leverage `standalone` components. This allows for more granular
dependency declaration and practically eliminates the need for
`NgModule` files.
- We've had to be more explicit with certain transitive dependencies,
for example `@sanity/asset-utils` with the new builder. But I believe
the benefit outweighs the cost. There should be a better solution here
since the website shouldn't really depend on `@sanity` modules that are
only used by the Sanity Studio anyway.