Pricing page + minimalist design iteration (#334)
## Release notes: usage and product changes

The Pricing page is now live!

We also patch severe rendering issues affecting our new topnav and
gradient buttons on Firefox and Safari.

## Implementation

The pricing page is based on the deployment page, but with significant
modifications to allow for longer tables and to show prices.

"View Details" takes you to the pricing details dialog, which breaks
down prices in detail and offers some sample configurations for each
provider for you to view the prices of.

Everything is mobile-friendly.

Other notable changes include the site banner becoming deliberately
hideous, and tables adopting a more flat, minimalist design.