Add edge function for authenticating the TypeDB Cloud API (#336)
## Release notes: usage and product changes

We add a function that accepts an API token client ID and client secret,
and exchanges them for an ID token from google identity platform.

## Implementation

The function extracts a client ID and client secret from a basic auth
header, and uses them to construct an appropriate request to google
identity platform. If the response is not `ok`, we return it as-is -
otherwise we extract the `idToken` and return that.

We specify separate API keys and tenant IDs for each environment in the
`netlify.toml`, to allow it to operate correctly for development,
staging, and production.