Update sanity schema to support conditional questions for surveys (#340)
## Release notes: usage and product changes

We update the sanity schema to support conditional questions - that is,
questions that may or may not be present depending on answers to other
questions.

## Implementation

The survey questions now have a new field, with the new type
`questionCondition`.

A `questionCondition` can:
- Be enabled/disabled entirely
- Switch between a show condition and a hide condition
- Switch between requiring *all* answers to match, or *any* answer to
match

It contains a list of matching answers, each of which consists of:
- The `posthogProperty` of the question to check
- A list of `posthogProperty`s for which that question will be
considered valid for the condition

We add a section-level validation function to ensure that all
`posthogProperty`s entered into conditions are in fact present in the
same section.