Introduce cloud announcement to sanity schema (#343)
## Release notes: usage and product changes

We add new schema for holding an announcement for TypeDB Cloud. This
announcement will be displayed on the overview page.

## Implementation

The cloud announcement consists of a title, subtitle, and link - which
will be displayed to the user; as well as an ID that is used to track
dismissal of the announcement, and a toggle to enable or disable the
announcement entirely.

We additionally export a sanity query used by TypeDB Cloud to retrieve
the announcement with the correctly formatted link.