Add main site sitemap (#346)
## Release notes: usage and product changes

We add a sitemap-main.xml generator script so that all our main site
pages get added into the sitemap with every deploy.

## Motivation

We forgot to do the sitemap when migrating from Scully to Angular SSR

## Implementation

We add the build-sitemap.js binary and invoke it on every run in the
`live` branch as well as the new `ssr` branch that we're adding to
enable testing SSR features without pushing to production.